requirements.txt
numpy<2
opencv-python>=4.5
ultralytics
torch
torchvision
Pillow
easyocr
PyQt6
pyobjc-framework-Quartz
pyobjc-core
